+++ apps/tlon-mobile/src/components/AuthenticatedApp.tsx
@@ -4,13 +4,16 @@ import {
 } from '@tloncorp/app/hooks/useAppStatusChange';
 import { useConfigureUrbitClient } from '@tloncorp/app/hooks/useConfigureUrbitClient';
 import { useFindSuggestedContacts } from '@tloncorp/app/hooks/useFindSuggestedContacts';
-import { useNavigationLogging } from '@tloncorp/app/hooks/useNavigationLogger';
 import { useNetworkLogger } from '@tloncorp/app/hooks/useNetworkLogger';
 import { useTelemetry } from '@tloncorp/app/hooks/useTelemetry';
 import { useUpdatePresentedNotifications } from '@tloncorp/app/lib/notifications';
 import { RootStack } from '@tloncorp/app/navigation/RootStack';
 import { AppDataProvider } from '@tloncorp/app/provider/AppDataProvider';
-import { PortalProvider, ZStack, useStore } from '@tloncorp/app/ui';
+import {
+  ForwardPostSheetProvider,
+  PortalProvider,
+  ZStack,
+} from '@tloncorp/app/ui';
 import { sync } from '@tloncorp/shared';
 import * as db from '@tloncorp/shared/db';
 import { useCallback, useEffect, useState } from 'react';
@@ -23,7 +26,6 @@ import { inviteSystemContacts } from '../lib/contactsHelpers';
 import { refreshHostingAuth } from '../lib/hostingAuth';
 
 function AuthenticatedApp() {
-  const store = useStore();
   const telemetry = useTelemetry();
   const checkNodeStopped = useCheckNodeStopped();
   useNotificationListener();
@@ -89,7 +91,11 @@ export default function ConnectedAuthenticatedApp() {
         This portal provider overrides the root portal provider 
         to ensure that sheets have access to `AppDataContext`
       */}
-      <PortalProvider>{clientReady && <AuthenticatedApp />}</PortalProvider>
+      <PortalProvider>
+        <ForwardPostSheetProvider>
+          {clientReady && <AuthenticatedApp />}
+        </ForwardPostSheetProvider>
+      </PortalProvider>
     </AppDataProvider>
   );
 }
