+++ packages/shared/src/types/ChannelActions.ts
@@ -10,7 +10,8 @@ export type Id =
   | 'edit'
   | 'report'
   | 'visibility'
-  | 'delete';
+  | 'delete'
+  | 'forward';
 
 /**
  * Info about a channel action type that should not change based on context.
@@ -34,6 +35,7 @@ export function channelActionIdsFor({
         'startThread',
         'muteThread',
         'copyRef',
+        'forward',
         'edit',
         'report',
         'visibility',
@@ -44,6 +46,7 @@ export function channelActionIdsFor({
         'startThread',
         'muteThread',
         'copyRef',
+        'forward',
         'edit',
         'report',
         'visibility',
@@ -67,6 +70,7 @@ export function channelActionIdsFor({
         'muteThread',
         'viewReactions',
         'copyRef',
+        'forward',
         'copyText',
         'edit',
         'visibility',
@@ -86,6 +90,7 @@ const STATIC_SPECS = {
   edit: { isNetworkDependent: true },
   muteThread: { isNetworkDependent: true },
   quote: { isNetworkDependent: true },
+  forward: { isNetworkDependent: true },
   report: { isNetworkDependent: true },
   startThread: { isNetworkDependent: true },
   viewReactions: { isNetworkDependent: false },
